requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coverage==5.5
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.3
django-allauth==0.44.0
django-braces==1.14.0
django-countries==7.2.1
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-model-utils==4.1.1
django-money==2.0
django-phonenumber-field==5.2.0
django-social-share==2.2.1
django-star-ratings==0.9.1
django-user-visit==0.4.1
geoip2==4.2.0
gunicorn==20.1.0
idna==2.10
maxminddb==2.0.3
multidict==5.1.0
oauthlib==3.1.1
phonenumberslite==8.12.24
Pillow==8.2.0
psycopg2==2.9
psycopg2-binary==2.9.1
py-moneyed==1.2
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
swapper==1.1.2.post1
typing-extensions==3.10.0.0
ua-parser==0.10.0
urllib3==1.26.5
user-agents==2.2.0
whitenoise==5.2.0
yarl==1.6.3
